Tweaked easing
By decreasing the initial xScale, I've made the transition of scaling
the side textures look more believable.

Previous awkwardness in spinning seemed to have stemmed from the fact
that a rotating cube is supposed to look wider when in mid rotation.
Experimented by scaling sideways by factor of 1.15. Seems to be a lot
more convincing now.